% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{BetaRandomVariable}
\alias{BetaRandomVariable}
\title{BetaRandomVariable object}
\description{
Defines a Beta random variable. It uses in the backend
all functions that defines a \code{\link[=rbeta]{Beta}} random variable. As
the other definitions in \code{distributions} this is an
\code{\link[R6:R6Class]{R6}} object. Tha parametrization used
is from a \deqn{X \sim \mathsf{Beta}(\alpha, \beta)\,,}
where the parameters \eqn{\alpha} and \eqn{\beta} are
the parameters defining a Beta random variable.
\subsection{Note}{

All random variables in \code{distributions} are defined as \code{\link{R6}}
objects. The \code{fields} (referenced below) are needed for objects of
\code{\link[R6:R6Class]{R6::R6Class}}. In our context these are what defines
which specific instance of a
random variable is used. That is, the public \code{fields} are the
parameters of the random variable \eqn{\theta \in \mathbb{R}^p}, for some
\eqn{p}.
}
}
\seealso{
Other continuous: 
\code{\link{GaussianRandomVariable}}
}
\concept{continuous}
\section{Super class}{
\code{distributions::ContinuousRandomVariable} -> \code{Beta}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{alpha}}{(\code{double}, positive) \cr
the first parameter of the distribution, \eqn{\alpha > 0}.}

\item{\code{beta}}{(\code{double}, positive) \cr
the second parameter of the distribution, \eqn{\beta > 0}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Beta-new}{\code{BetaRandomVariable$new()}}
\item \href{#method-Beta-sample}{\code{BetaRandomVariable$sample()}}
\item \href{#method-Beta-density}{\code{BetaRandomVariable$density()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Beta-new"></a>}}
\if{latex}{\out{\hypertarget{method-Beta-new}{}}}
\subsection{Method \code{new()}}{
Generates a Beta \code{\link{ContinuousRandomVariable}} object with specified
parameters alpha (\code{alpha}) and beta (\code{beta}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaRandomVariable$new(alpha = 1, beta = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{alpha}}{(\code{double}, positive) the first parameter.}

\item{\code{beta}}{(\code{double}, positive) the second parameter.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Beta-sample"></a>}}
\if{latex}{\out{\hypertarget{method-Beta-sample}{}}}
\subsection{Method \code{sample()}}{
Generates random variables using the \code{\link[stats:Beta]{stats::rbeta()}} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaRandomVariable$sample(nsamples = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nsamples}}{(integer) the number of random numbers
being generated that behaves like the random variables.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{array}) of random numbers from the distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Beta-density"></a>}}
\if{latex}{\out{\hypertarget{method-Beta-density}{}}}
\subsection{Method \code{density()}}{
Evaluates the density function using the \code{\link[stats:Beta]{stats::dbeta()}} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BetaRandomVariable$density(x, log = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(array) to evaluate the density function.}

\item{\code{log}}{(Boolean) indicates if we are evaluating the \code{log-density}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{array}) of evaluations at \code{x}.
}
}
}
