% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{BionomialRandomVariable}
\alias{BionomialRandomVariable}
\alias{BinomialRandomVariable}
\title{BinomialRandomVariable object}
\description{
Defines a Binomial random variable. It uses in the backend
all functions that defines a \code{\link[=rbinom]{Binomial}} random variable. As
the other definitions in \code{distributions} this is an
\code{\link[R6:R6Class]{R6}} object. Tha parametrization used
is from a \deqn{X \sim \mathsf{Binomial}(n, \theta)\,,}
where the parameters \eqn{n} and \eqn{\theta} are
the number of independent trials and the rate (probability) of success.
\subsection{Note}{

All random variables in \code{distributions} are defined as \code{\link{R6}}
objects. The \code{fields} (referenced below) are needed for objects of
\code{\link[R6:R6Class]{R6::R6Class}}. In our context these are what defines
which specific instance of a
random variable is used. That is, the public \code{fields} are the
parameters of the random variable \eqn{\theta \in \mathbb{R}^p}, for some
\eqn{p}.
}
}
\seealso{
Other discrete: 
\code{\link{PoissonRandomVariable}}
}
\concept{discrete}
\section{Super class}{
\code{distributions::DiscreteRandomVariable} -> \code{Binomial}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{n}}{(\code{integer}) \cr
the number of independent trials, \eqn{n}.}

\item{\code{theta}}{(\code{double}, constrained) \cr
the probability of success of an independent trial,
\eqn{\theta \in [0,1]}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Binomial-new}{\code{BinomialRandomVariable$new()}}
\item \href{#method-Binomial-sample}{\code{BinomialRandomVariable$sample()}}
\item \href{#method-Binomial-density}{\code{BinomialRandomVariable$density()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Binomial-new"></a>}}
\if{latex}{\out{\hypertarget{method-Binomial-new}{}}}
\subsection{Method \code{new()}}{
Generates a Binomial \code{\link{DiscreteRandomVariable}} object with specified
number of trials (\code{n}) and probability of success (\code{theta}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinomialRandomVariable$new(n = 0, theta = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{(\code{integer}) the center of the distribution.}

\item{\code{theta}}{(\code{double}, constrained) the dispersion around
the center of the distribution.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Binomial-sample"></a>}}
\if{latex}{\out{\hypertarget{method-Binomial-sample}{}}}
\subsection{Method \code{sample()}}{
Generates random variables using the \code{\link[stats:Binomial]{stats::rbinom()}} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinomialRandomVariable$sample(nsamples = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nsamples}}{(\code{integer}) the number of random numbers
being generated that behaves like the random variables.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{array}) of random numbers from the distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Binomial-density"></a>}}
\if{latex}{\out{\hypertarget{method-Binomial-density}{}}}
\subsection{Method \code{density()}}{
Evaluates the density function using the \code{\link[stats:Binomial]{stats::dbinom()}} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{BinomialRandomVariable$density(x, log = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(array) to evaluate the density function.}

\item{\code{log}}{(Boolean) indicates if we are evaluating the \code{log-density}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{array}) of evaluations at \code{x}.
}
}
}
