% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{calculate.cdf}
\alias{calculate.cdf}
\title{Calculate the cumulative distribution function for a distribution}
\usage{
calculate.cdf(dist, q, lower.tail = T, log.p = F, n.sim = 1000 * dist@n.var)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{n.sim}{If the values are to be calculated by random sample generation, how many random samples to generate}

\item{x}{The values of the parameters at which to calculate the CDF. May be either a vector, if the CDF at a single point is desired, or a matrix where each row represents one point and each column represents a variable in the distribution}

\item{log}{A logical indicating whether to return the CDF on the log scale}
}
\value{
If x is a vector, a scalar density. If x is a matrix, a numeric vector of densities, one for each row in x
}
\description{
Calculate the cumulative distribution function for a distribution
}
\details{
Note to developers: this function should not be overridden. Override \code{\link{do.calculate.cdf}} instead
}
