% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{calculate.density}
\alias{calculate.density}
\title{Calculate the density for a distribution}
\usage{
calculate.density(dist, x, log = F, n.sim = 1000)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{x}{The values of the parameters at which to calculate the density. May be either a vector, if the density at a single point is desired, or a matrix where each row represents one point and each column represents a variable in the distribution}

\item{log}{A logical indicating whether to return the density on the log scale}

\item{n.sim}{If the values are to be calculated by random sample generation, how many random samples to generate}
}
\value{
If x is a vector, a scalar density. If x is a matrix, a numeric vector of densities, one for each row in x
}
\description{
Calculate the density for a distribution
}
\details{
Note to developers: This function should not be overridden directly. Insteady override \code{\link{do.calculate.density}}
}
\keyword{internal}
