% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{calculate.marginal.cdfs}
\alias{calculate.marginal.cdfs}
\title{Calculate a marginal cumulative distribution function for each variable in a distribution}
\usage{
calculate.marginal.cdfs(dist, q, lower.tail = T, log.p = F, n.sim = 1000)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{n.sim}{If the values are to be calculated by random sample generation, how many random samples to generate}
}
\value{
If q is a vector, a vector of probabilities with one value for each variable in the distribution. If q is a matrix, a matrix of probabilites, with one for each row in x and one column for each variable in the distribution
}
\description{
Calculate a marginal cumulative distribution function for each variable in a distribution
}
\details{
Note to developers: this function should not be overridden. Override \code{\link{do.calculate.marginal.cdfs}} instead
}
