% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{calculate.marginal.densities}
\alias{calculate.marginal.densities}
\title{Calculate a marginal density for each variable in a distribution}
\usage{
calculate.marginal.densities(dist, x, log = F, n.sim = 1000)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{x}{The values of the parameters at which to calculate the density. May be either a vector, if the density at a single point is desired, or a matrix where each row represents one point and each column represents a variable in the distribution}

\item{log}{A logical indicating whether to return the density on the log scale}

\item{n.sim}{If the values are to be calculated by random sample generation, how many random samples to generate}
}
\value{
If x is a vector or a matrix with only one row, a vector of densities with one value for each variable in the distribution. If x is a matrix, a matrix of densities, with one for each row in x and one column for each variable in the distribution
}
\description{
Calculate a marginal density for each variable in a distribution
}
\details{
Note to developers: this function should not be overridden. Override \code{\link{do.calculate.marginal.densities}} instead
}
