% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_mixture_distribution.R
\name{Canonical.Mixture.Distribution}
\alias{Canonical.Mixture.Distribution}
\title{Create a mixture distribution object based off of a canonical, univariate distribution}
\usage{
Canonical.Mixture.Distribution(
  name,
  dist.name,
  parameters,
  weights = 1,
  var.name = NULL,
  support = Continuous.Support(),
  density.function.name = paste0("d", dist.name),
  cdf.function.name = paste0("p", dist.name),
  quantile.function.name = paste0("q", dist.name),
  transformation = NULL,
  is.improper = F,
  mean.values = as.numeric(NA),
  variance.values = as.numeric(NA)
)
}
\arguments{
\item{name}{A short, descriptive name for the distribution}

\item{dist.name}{The root name of the distribution as used by d/r/p functions in R. For example, to create a normal distribution, dist.name='norm' (like dnorm, rnorm, pnorm). To create a beta distribution, dist.name='beta' (dbeta, rbeta, qbeta). This argument is only necessary if density.function, cdf.function, and quantile.function are not passed explicitly}

\item{parameters}{A list of vectors, where each element has parameter values that would be passed to d/r/p functions in R. For example, for a mixture of normal(0,1) and normal(1,1) distributions, parameters=list(mean=c(0,1), sd=1) or parameters=list(mean=c(0,1), sd=c(1,1))}

\item{weights}{A weight for each component distribution in the mixture. The weights need not sum to 1. If a scalar is passed, all components are given the same weight}

\item{var.name}{The name of the variable in this distribution}

\item{transformation}{A transformation object, if the named distribution operates on a transformation of the random variable (can also pass NULL for no transformation or the name of a predefined transformation - see \code{\link{get.defined.transformation}}). For example, to create a log-normal distribution, use dist.name='norm' with transformation='log}

\item{is.improper, is.discrete}{Logicals indicating whether the distribution is improper/discrete}

\item{mean.values, variance.values}{If known, vectors of values for the means and variances of each component of the distribution}

\item{lower.bound, upper.bound}{If this is a bounded distribution, the lower and upper bounds}

\item{density.function, cdf.function, quantile.function}{The functions that calculate the density, cdf, and quantile respectively. These should take the standard parameters of such functions in R. density.function should take 'x', distribution-specific parameters, and 'log'. cdf.function should take 'q', distribution-specific parameters, 'lower.tail' and 'log.p'. quantile.function should take 'p', distribution-specific parameters, and 'lower.tail' and 'log.p'}
}
\description{

}
\seealso{
\code{\link{get.defined.transformation}}, \code{\link{create.transformation}}

Other Univariate Canonical Distribution Constructors: 
\code{\link{Bernoulli.Distribution}()},
\code{\link{Beta.Distribution}()},
\code{\link{Binomial.Distribution}()},
\code{\link{Logitnormal.Distribution}()},
\code{\link{Logituniform.Distribution}()},
\code{\link{Lognormal.Distribution}()},
\code{\link{Loguniform.Distribution}()},
\code{\link{Normal.Distribution}()},
\code{\link{Transformed.Normal.Distribution}()},
\code{\link{Uniform.Distribution}()},
\code{\link{Univariate.Canonical.Distribution}()}

Other Distribution Constructors: 
\code{\link{Autoregressive.Multivariate.Normal.Distribution}()},
\code{\link{Bernoulli.Distribution}()},
\code{\link{Beta.Distribution}()},
\code{\link{Binomial.Distribution}()},
\code{\link{Compound.Symmetry.Multivariate.Normal.Distribution}()},
\code{\link{Empiric.Distribution}()},
\code{\link{Logitnormal.Distribution}()},
\code{\link{Logitnormal.Mixture}()},
\code{\link{Logituniform.Distribution}()},
\code{\link{Lognormal.Distribution}()},
\code{\link{Lognormal.Mixture}()},
\code{\link{Loguniform.Distribution}()},
\code{\link{Multivariate.Correlated.Uniform.Distribution}()},
\code{\link{Multivariate.Logitnormal.Distribution}()},
\code{\link{Multivariate.Lognormal.Distribution}()},
\code{\link{Multivariate.Normal.Distribution}()},
\code{\link{Normal.Distribution}()},
\code{\link{Normal.Mixture}()},
\code{\link{Smoothed.Empiric.Distribution}()},
\code{\link{Transformed.Multivariate.Normal.Distribution}()},
\code{\link{Transformed.Normal.Distribution}()},
\code{\link{Transformed.Normal.Mixture}()},
\code{\link{Uniform.Distribution}()},
\code{\link{Univariate.Canonical.Distribution}()},
\code{\link{join.distributions}()}
}
\concept{Distribution Constructors}
\concept{Univariate Canonical Distribution Constructors}
