% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_mixture_distribution.R
\docType{class}
\name{Canonical_Mixture_Distribution}
\alias{Canonical_Mixture_Distribution}
\alias{Canonical_Mixture_Distribution-class}
\title{The Canonical_Mixture_Distribution class}
\description{
A class that represents a univariate mixture of different instances of a canonical distribution (the same distribution, but with different parameters)
}
\section{Slots}{

\describe{
\item{\code{name}}{A short, descriptive name for the distribution}

\item{\code{n.components}}{The number of component distributions in this mixture}

\item{\code{weights}}{The weights for the component distributions in the mixture (sum(weights)=1)}

\item{\code{density.function.name,cdf.function.name,quantile.function.name}}{Names of the (elsewhere defined) functions that calculate density, cdf, and quantiles (eg 'dnorm', 'pnorm', and 'qnorm)}

\item{\code{parameters}}{A list of vectors; each element of the list represents a parameter to the distribution, and the values of the vector represent the parameter values for each component of the mixture}

\item{\code{p.in.bounds,log.p.in.bounds}}{A vector cumulative probabilities and of the log cumulative probabilities between the lower and upper bounds for each component of the mixture}

\item{\code{p.lower.bound}}{A vector of values of the cdf evaluated at the lower bound for each components of the distribution}

\item{\code{transformation}}{An object of class \link{transformation}, defining the transformation applied to the variable before applying the distribution (eg, )}

\item{\code{means,variances}}{If known, the means and variances of each component of the distribution. May be NA}
}}

\seealso{
The family of functions that create specific mixture distributions, such as \link{normal.mixture}
}
