% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_constructors.R
\name{Compound.Symmetry.Multivariate.Normal.Distribution}
\alias{Compound.Symmetry.Multivariate.Normal.Distribution}
\title{Create a multivariate normal distribution with a Compound Symmetry correlation structure}
\usage{
Compound.Symmetry.Multivariate.Normal.Distribution(
  rho,
  n = max(length(mu), length(sds)),
  mu = 0,
  sds = 1,
  lower = -Inf,
  upper = Inf,
  var.names = NULL,
  transformations = NULL
)
}
\arguments{
\item{rho}{The correlation coefficient}

\item{n}{The number of variables in the distribution}

\item{mu, sds}{The mean and standard deviation vectors. If given a scalar value, assumes that the mean/sd is the same for every variable in the distribution}

\item{lower}{The upper and lower bounds for each variable (on the original scale, NOT the transformed scale). If only one value is given, assumed to apply to all variables in the distribution}

\item{upper}{The upper and lower bounds for each variable (on the original scale, NOT the transformed scale). If only one value is given, assumed to apply to all variables in the distribution}

\item{var.names}{The names of the variables in the distribution. If NULL, uses the names of mu, the row names of sigma, or the column names of sigma, in that order}

\item{transformations}{Objects of class \link{transformation} to be applied to each variable. Can be either an object of class \link{transformation}, the character name of a defined transformation, or a list of \link{transformation} objects.   If only one transformation is supplied, assumed to apply to all variables. If this is a named list, any variables missing transformation are assmed to not be transformed}
}
\description{
Creates a distribution where all the variables, after arbitrary transformations, follow a normal distribution with a Compound Symmetry (AKA Exchangeable) correlation structure
}
\seealso{
Other Distribution Constructors: 
\code{\link{Autoregressive.Multivariate.Normal.Distribution}()},
\code{\link{Bernoulli.Distribution}()},
\code{\link{Beta.Distribution}()},
\code{\link{Binomial.Distribution}()},
\code{\link{Canonical.Mixture.Distribution}()},
\code{\link{Empiric.Distribution}()},
\code{\link{Logitnormal.Distribution}()},
\code{\link{Logitnormal.Mixture}()},
\code{\link{Logituniform.Distribution}()},
\code{\link{Lognormal.Distribution}()},
\code{\link{Lognormal.Mixture}()},
\code{\link{Loguniform.Distribution}()},
\code{\link{Multivariate.Logitnormal.Distribution}()},
\code{\link{Multivariate.Lognormal.Distribution}()},
\code{\link{Multivariate.Normal.Distribution}()},
\code{\link{Normal.Distribution}()},
\code{\link{Normal.Mixture}()},
\code{\link{Smoothed.Empiric.Distribution}()},
\code{\link{Transformed.Multivariate.Normal.Distribution}()},
\code{\link{Transformed.Normal.Distribution}()},
\code{\link{Transformed.Normal.Mixture}()},
\code{\link{Uniform.Distribution}()},
\code{\link{Univariate.Canonical.Distribution}()},
\code{\link{join.distributions}()}
}
\concept{Distribution Constructors}
