% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_set_distributions.R
\name{Constant.Distribution}
\alias{Constant.Distribution}
\title{Create a distribution representing a Constant Distribution}
\usage{
Constant.Distribution(value, var.name = NULL)
}
\arguments{
\item{value}{The single value supported by the distribution}

\item{var.name}{The name of the single variable in the distribution, or NULL if no name is specified}
}
\description{
Creates an object representing a degenerate distribution that only supports a single, discrete value
}
\seealso{
\link{Discrete.Set.Distribution}

Other Discrete Distribution Constructors: 
\code{\link{Discrete.Set.Distribution}()}

Other Distribution Constructors: 
\code{\link{Autoregressive.Multivariate.Normal.Distribution}()},
\code{\link{Bernoulli.Distribution}()},
\code{\link{Beta.Distribution}()},
\code{\link{Binomial.Distribution}()},
\code{\link{Canonical.Mixture.Distribution}()},
\code{\link{Compound.Symmetry.Multivariate.Normal.Distribution}()},
\code{\link{Discrete.Set.Distribution}()},
\code{\link{Empiric.Distribution}()},
\code{\link{Logitnormal.Distribution}()},
\code{\link{Logitnormal.Mixture}()},
\code{\link{Logituniform.Distribution}()},
\code{\link{Lognormal.Distribution}()},
\code{\link{Lognormal.Mixture}()},
\code{\link{Loguniform.Distribution}()},
\code{\link{Multivariate.Correlated.Uniform.Distribution}()},
\code{\link{Multivariate.Logitnormal.Distribution}()},
\code{\link{Multivariate.Lognormal.Distribution}()},
\code{\link{Multivariate.Normal.Distribution}()},
\code{\link{Normal.Distribution}()},
\code{\link{Normal.Mixture}()},
\code{\link{Smoothed.Empiric.Distribution}()},
\code{\link{Transformed.Multivariate.Normal.Distribution}()},
\code{\link{Transformed.Normal.Distribution}()},
\code{\link{Transformed.Normal.Mixture}()},
\code{\link{Uniform.Distribution}()},
\code{\link{Univariate.Canonical.Distribution}()},
\code{\link{join.distributions}()}
}
\concept{Discrete Distribution Constructors}
\concept{Distribution Constructors}
