% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_set_distributions.R
\docType{class}
\name{Constant_Distribution}
\alias{Constant_Distribution}
\alias{Constant_Distribution-class}
\title{The Constant_Distribution class}
\description{
A class that represents a probility distribution with support for exactly one value
}
\section{Slots}{

\describe{
\item{\code{value}}{The single value supported by this distribution}
}}

\seealso{
The \link{Discrete.Set.Distribution}
}
