% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{Continuous.Support}
\alias{Continuous.Support}
\title{Create an object which represents support over a range of integers}
\usage{
Continuous.Support(
  lower = -Inf,
  upper = Inf,
  lower.inclusive = F,
  upper.inclusive = F
)
}
\arguments{
\item{lower, upper}{The lower and upper bounds of support. For a single range of support, these should both be of length 1. If support exists over multiple disjoint ranges, then lower and upper should be numeric vectors of the same length, with the lower and upper bounds of each of the disjoint ranges}

\item{lower.inclusive, upper.inclusive}{Logical vectors indicating whether the lower and upper bounds are inclusive (vs. exclusive). If length(lower)/length(upper) are >1, these can be either logical vectors of the same length, or logical scalars (in which case the value applies to each element of lower and upper)}
}
\value{
An object of class \link{Range_Support}, such that is.supported is true if x falls within any of the ranges lower[i] to upper[i]
}
\description{
Create an object which represents support over a range of integers
}
