% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{ContinuousRandomVariable}
\alias{ContinuousRandomVariable}
\title{Continuous Random variable structure}
\value{
A \code{\link{RandomVariable}} object.
}
\description{
Sets the structure of random variables objects in the
continuous setting
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ContinuousRandomVariable-samples}{\code{ContinuousRandomVariable$samples()}}
\item \href{#method-ContinuousRandomVariable-density}{\code{ContinuousRandomVariable$density()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContinuousRandomVariable-samples"></a>}}
\if{latex}{\out{\hypertarget{method-ContinuousRandomVariable-samples}{}}}
\subsection{Method \code{samples()}}{
Generates random numbers from that behave like
realizations of the random variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContinuousRandomVariable$samples(nsamples = 1, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nsamples}}{number of samples to generate.}

\item{\code{...}}{allows for more inputs.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ContinuousRandomVariable-density"></a>}}
\if{latex}{\out{\hypertarget{method-ContinuousRandomVariable-density}{}}}
\subsection{Method \code{density()}}{
Evaluates the \code{pdf} of the random variable.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ContinuousRandomVariable$density(x, log = TRUE, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{\code{\link{array(double)}} input on where to evaluate density function.}

\item{\code{log}}{\code{\link{Boolean}} indicates if we need the log-density.}

\item{\code{...}}{allows for more inputs.}
}
\if{html}{\out{</div>}}
}
}
}
