% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{create.transformation}
\alias{create.transformation}
\title{Create an object representing a transformation}
\usage{
create.transformation(
  transform.function,
  reverse.transform.function,
  transformation.derivative,
  log.abs.transformation.derivative = NULL,
  name
)
}
\arguments{
\item{transform.function, reverse.transform.function, transformation.derivative, log.transformation.derivative}{Functions that compute the transformation, inverse transformation, derivative of the transformation function, and log of the derivative of the transformation function, directly. All should take one argument, 'x', which can be either a numeric vector or numeric scalar. log.transformation.derivative may be supplied as NULL; the other three functions must be defined}

\item{name}{A character value with a descriptive name of the transformation}
}
\description{
Create an object representing a transformation
}
