\name{cugen}
\alias{cugen}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Uniform Numbers
}
\description{
Cugen function use "dugen" to create uniform random number between 0 and 1
}

\usage{
cugen()
}
\examples{
cugen()
}
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
