% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_set_distributions.R
\name{Discrete.Set.Distribution}
\alias{Discrete.Set.Distribution}
\title{Create a distribution representing a Discrete Set}
\usage{
Discrete.Set.Distribution(values, weights, var.name = NULL)
}
\arguments{
\item{values}{The numeric set of values}

\item{weights}{The weight given to the corresponding value. Need not sum to 1}

\item{var.name}{The name of the single variable in the distribution, or NULL if no name is specified}
}
\description{
Creates an object representing a discrete probability distribution over an arbitrary set of numbers
}
\seealso{
\link{Constant.Distribution}

Other Discrete Distribution Constructors: 
\code{\link{Constant.Distribution}()}

Other Distribution Constructors: 
\code{\link{Autoregressive.Multivariate.Normal.Distribution}()},
\code{\link{Bernoulli.Distribution}()},
\code{\link{Beta.Distribution}()},
\code{\link{Binomial.Distribution}()},
\code{\link{Canonical.Mixture.Distribution}()},
\code{\link{Compound.Symmetry.Multivariate.Normal.Distribution}()},
\code{\link{Constant.Distribution}()},
\code{\link{Empiric.Distribution}()},
\code{\link{Logitnormal.Distribution}()},
\code{\link{Logitnormal.Mixture}()},
\code{\link{Logituniform.Distribution}()},
\code{\link{Lognormal.Distribution}()},
\code{\link{Lognormal.Mixture}()},
\code{\link{Loguniform.Distribution}()},
\code{\link{Multivariate.Correlated.Uniform.Distribution}()},
\code{\link{Multivariate.Logitnormal.Distribution}()},
\code{\link{Multivariate.Lognormal.Distribution}()},
\code{\link{Multivariate.Normal.Distribution}()},
\code{\link{Normal.Distribution}()},
\code{\link{Normal.Mixture}()},
\code{\link{Smoothed.Empiric.Distribution}()},
\code{\link{Transformed.Multivariate.Normal.Distribution}()},
\code{\link{Transformed.Normal.Distribution}()},
\code{\link{Transformed.Normal.Mixture}()},
\code{\link{Uniform.Distribution}()},
\code{\link{Univariate.Canonical.Distribution}()},
\code{\link{join.distributions}()}
}
\concept{Discrete Distribution Constructors}
\concept{Distribution Constructors}
