% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{Discrete.Set.Support}
\alias{Discrete.Set.Support}
\title{Create a support over an arbitrary set of values}
\usage{
Discrete.Set.Support(supported.values)
}
\arguments{
\item{supported.values}{A numeric vector of the allowed values in this support}
}
\value{
An object of class \link{Discrete_Set_Support}
}
\description{
Create a support over an arbitrary set of values
}
