% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_set_distributions.R
\docType{class}
\name{Discrete_Set_Distribution}
\alias{Discrete_Set_Distribution}
\alias{Discrete_Set_Distribution-class}
\title{The Discrete_Set_Distribution class}
\description{
A class that represents a discrete probability distribution over an arbitrary set of numbers
}
\section{Slots}{

\describe{
\item{\code{values}}{The numeric set of values}

\item{\code{weights}}{The probability of each value in the set. Sum to 1}
}}

\seealso{
The \link{Constant.Distribution}
}
