% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\docType{class}
\name{Discrete_Set_Support}
\alias{Discrete_Set_Support}
\alias{Discrete_Set_Support-class}
\title{An object representing support over an arbitrary set of real numbers}
\description{
An object representing support over an arbitrary set of real numbers
}
