% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\docType{class}
\name{Distribution}
\alias{Distribution}
\alias{Distribution-class}
\title{The abstract superclass for all Distribution objects}
\description{
Distribution objects represent a distribution of one or more variables
}
\section{Slots}{

\describe{
\item{\code{var.names}}{A character vector of the names of the variables in the distribution (optional)}

\item{\code{n.var}}{The number of variables in the distribution}

\item{\code{support}}{A \link{Support} object defining the support for each variable in the distribution}

\item{\code{is.improper}}{A logical vectors indicating whether each variable in the distribution is improper}
}}

\seealso{
Specific subclasses which can be instantiated:
\itemize{
    \item{\link{Empiric_Distribution}}
    \item{\link{Univariate_Canonical_Distribution}}
    \item{\link{Joint_Independent_Distributions}}
}
}
