\name{distributions-package}
\alias{distributions-package}
\docType{package}
\title{
Probability distributions based on TI-83 Plus
}
\description{
Probability distributions (binomial, poisson, geometric,
normal, chi square, Fisher, Student) based on TI-83 Plus
graphic scientific calculator
}
\details{
\tabular{ll}{
Package: \tab distributions\cr
Type: \tab Package\cr
Encoding: \tab latin1\cr
Depends: \tab R, stats, utils, tools\cr
Version: \tab 1.3\cr
Date: \tab 2007-10-12\cr
License: \tab GPL-2\cr
}

Index:
\preformatted{
Fcdf                    Distribuzione di probabilità cumulata F
Fpdf                    Densità di probabilità F
binomcdf                Densità binomiale cumulata
binompdf                Probabilità binomiale
chi2cdf                 Distribuzione cumulata della probabilità chi
                        quadrato
chi2pdf                 Densità di probabilità chi quadrato
geometcdf               Densità cumulata geometrica
geometpdf               Probabilità geometrica
invNorm                 Distribuzione cumulata normale inversa
normalcdf               Distribuzione cumulata della probabilità
                        normale
normalpdf               Densità di probabilità normale
poissoncdf              Densità cumulata di Poisson
poissonpdf              Probabilità di Poisson
tcdf                    Ditribuzione della probabilità t di Student
tpdf                    Densità di probabilità t di Student
}
}
\author{
Fabio Frascati

Maintainer: Fabio Frascati <fabiofrascati@yahoo.it>
}
\keyword{ package }
