% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{distributions.equal}
\alias{distributions.equal}
\title{Tests whether two Distribution objects represent the same distribution}
\usage{
distributions.equal(dist1, dist2)
}
\arguments{
\item{dist1, dist2}{The distributions to compare}
}
\value{
True if the two distributions produce the same random samples
}
\description{
Evaluates to true if the two distributions produce the same random samples
}
\details{
The variable order must be the same for two multivariate distributions, otherwise the equality test will return false
}
