% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{do.calculate.cdf}
\alias{do.calculate.cdf}
\title{Internal function for subclasses to calculate the cumulative density function for a distribution}
\usage{
do.calculate.cdf(dist, q, lower.tail = T, log.p = F, n.sim = 1000 * dist@n.var)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{n.sim}{If the values are to be calculated by random sample generation, how many random samples to generate}
}
\description{
This function is provided to be overriden by developers. End users should instead call \code{\link{calculate.cdf}}
}
\details{
The default implementation generates random samples and takes the mean number of samples where all variables are <= the given values of q
The wrapper \code{\link{calculate.cdf}} takes care of matching variables (so the 'q' passed to this function is clean) and formatting the return value and generating warnings for any improper variables
}
\keyword{internal}
