% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{do.calculate.density}
\alias{do.calculate.density}
\title{Internal function for subclasses to calculate the density for a distribution}
\usage{
do.calculate.density(dist, x, log = F, n.sim = 1000)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{x}{A matrix where each row represents one point and each column represents a variable in the distribution (the wrapper matches variable names before passing to this function)}

\item{log}{A logical indicating whether to return the density on the log scale}

\item{n.sim}{If the values are to be calculated by random sample generation, how many random samples to generate}
}
\value{
If x is a vector, a scalar density. If x is a matrix, a numeric vector of densities, one for each row in x
}
\description{
This function is provided to be overriden by developers. End users should instead call \code{\link{calculate.density}}
}
\details{
The wrapper \code{\link{calculate.density}} takes care of matching variable names and of formatting the return value
}
