% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{do.calculate.marginal.cdfs}
\alias{do.calculate.marginal.cdfs}
\title{Internal function for subclasses tocalculate a marginal cumulative distribution function for each variable in a distribution}
\usage{
do.calculate.marginal.cdfs(dist, q, lower.tail = T, log.p = F, n.sim = 1000)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{n.sim}{If the values are to be calculated by random sample generation, how many random samples to generate}
}
\description{
This function is provided to be overriden by developers. End users should instead call \code{\link{calculate.marginal.cdfs}}
}
\details{
By default, implemented only for distributions with only one variable
The wrapper \code{\link{calculate.marginal.cdfs}} takes care of matching variables (so the 'q' passed to this function is clean) and formatting the return value and generating warnings for any improper variables
}
\keyword{internal}
