% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{do.calculate.marginal.densities}
\alias{do.calculate.marginal.densities}
\title{An internal function for subclasses to calculate a marginal density for each variable in a distribution}
\usage{
do.calculate.marginal.densities(dist, x, log = F, n.sim = 1000)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{x}{The values of the parameters at which to calculate the density. May be either a vector, if the density at a single point is desired, or a matrix where each row represents one point and each column represents a variable in the distribution}

\item{log}{A logical indicating whether to return the density on the log scale}

\item{n.sim}{If the values are to be calculated by random sample generation, how many random samples to generate}
}
\description{
This function is provided to be overriden by developers. End users should instead call \code{\link{calculate.marginal.densities}}
}
\details{
By default, implemented only for distributions with only one variable
The wrapper \code{\link{calculate.marginal.densities}} takes care of matching variables (so the 'x' passed to this function is clean) and formatting the return value
}
\keyword{internal}
