% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{do.generate.random.marginal.samples}
\alias{do.generate.random.marginal.samples}
\title{Internal function for subclasses to generate marginal random samples for each variable in a distribution}
\usage{
do.generate.random.marginal.samples(dist, n)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{n}{The number of random samples to be generated}
}
\description{
This function is provided to be overriden by developers. End users should instead call \code{\link{generate.random.marginal.samples}}
}
\details{
The default implementation just calls the standard \link{do.generate.random.samples}
The wrapper \code{\link{generate.random.marginal.samples}} takes care of formatting the return value and generating warnings for any improper variables
}
\keyword{internal}
