% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{do.generate.random.samples}
\alias{do.generate.random.samples}
\title{Internal function for subclasses to generate random samples from a distribution}
\usage{
do.generate.random.samples(dist, n)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{n}{The number of random samples to be generated}
}
\value{
If n==1, a numeric vector with one value for each variable in the distribution. If there is only one variable in the distribution, a numeric vector with n values. If n>1 and there is more than one variable in the distribution, a matrix with n rows, where each row represents one random sample, and each column represents a variable in the distribution
}
\description{
This function is provided to be overriden by developers. End users should instead call \code{\link{generate.random.samples}}
}
\details{
The wrapper \code{\link{generate.random.samples}} takes care of formatting the return value and generating warnings for any improper variables
}
\keyword{internal}
