% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{do.get.covariance.matrix}
\alias{do.get.covariance.matrix}
\title{Internal function for subclasses to get the variance-covariance matrix for variables in a distribution}
\usage{
do.get.covariance.matrix(dist, n.sim)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{n.sim}{If the values are to be calculated by random sample generation, how many random samples to generate}
}
\description{
This function is provided to be overriden by developers. End users should instead call \code{\link{get.covariance.matrix}}
}
\details{
The default implementation generates random samples and takes the sample covariance.
The wrapper \code{\link{get.covariance.matrix}} takes care of formatting the return value and generating warnings for any improper variables
}
\keyword{internal}
