% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{do.get.means}
\alias{do.get.means}
\title{Internal function for subclasses to get means of the variables in a distribution}
\usage{
do.get.means(dist, n.sim = 1000)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{n.sim}{If the values are to be calculated by random sample generation, how many random samples to generate}
}
\value{
A numeric vector with a mean for each variable in the distribution
}
\description{
This function is provided to be overriden by developers. End users should instead call \code{\link{get.means}}
}
\details{
The default implementation generates random samples and takes the sample mean for each variable.
The wrapper \code{\link{get.means}} takes care of formatting the return value and generating warnings for any improper variables
}
\keyword{internal}
