% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{do.get.medians}
\alias{do.get.medians}
\title{Internal function for subclasses to get medians of the variables in a distribution}
\usage{
do.get.medians(dist, n.sim)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{n.sim}{If the values are to be calculated by random sample generation, how many random samples to generate}
}
\value{
A numeric vector with a median for each variable in the distribution
}
\description{
This function is provided to be overriden by developers. End users should instead call \code{\link{get.medians}}
}
\details{
The wrapper \code{\link{get.medians}} takes care of formatting the return value and generating warnings for any improper variables
}
\keyword{internal}
