% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{do.get.sds}
\alias{do.get.sds}
\title{An internal function for subclasses to get the standard deviations of the variables in a distribution}
\usage{
do.get.sds(dist, n.sim = 1000)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{n.sim}{If the values are to be calculated by random sample generation, how many random samples to generate}
}
\description{
This function is provided to be overriden by developers. End users should instead call \code{\link{get.sds}}
}
\details{
The default implementation calls \code{\link{get.covariance.matrix}} and takes the square-root of the diagonal
The wrapper \code{\link{get.sds}} takes care of formatting the return value and generating warnings for any improper variables
}
\keyword{internal}
