% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{do.subset.distribution}
\alias{do.subset.distribution}
\title{Internal function for subclasses to get a distribution that describes a subset of the variables in a given distribution}
\usage{
do.subset.distribution(dist, keep.indices)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{keep.indices}{The numeric indices of which variables of the distribution (in the order they appear in var.names) to keep (this will be a subset of 1:dist@n.var)}
}
\description{
This function is provided to be overriden by developers. End users should instead call \code{\link{subset.distribution}}
}
\details{
Not implemented by default
The wrapper \code{\link{subset.distribution}} takes care of matching variables to their numeric indices. This function is only called by the wrapper if a proper subset (ie, not all the variables) is desired
}
\keyword{internal}
