\name{dugen}
\alias{dugen}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Uniform Numbers
}
\description{
Dugen function use "rgenerator" to create uniform random number between two integer number , these integers are taking as input
}
\usage{
dugen(min , max)
}
\arguments{
  \item{min}{
    \code{
    minimum
    }
  }
    \item{max}{
    \code{
    maximum
    }
  }
}
\examples{
dugen(3,8)
}
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
