% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empiric_distribution.R
\docType{class}
\name{Empiric_Distribution}
\alias{Empiric_Distribution}
\alias{Empiric_Distribution-class}
\title{The Empiric_Distribution Class}
\description{
A class that defines a distribution based on a set of samples from that distribution. Treats the samples as coming from a discrete distribution. For a continuous representation, see \link{Smoothed_Empiric_Distribution}
}
\section{Slots}{

\describe{
\item{\code{points}}{A matrix containing the samples}

\item{\code{weights}}{A numeric vector with the weight given to each row in points. Sum to 1}

\item{\code{n.points}}{The number of rows in points}
}}

\seealso{
\link{Smoothed_Empiric_Distribution}, \link{create.empiric.distribution}
}
