% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{GaussianRandomVariable}
\alias{GaussianRandomVariable}
\title{GaussianRandomVariable object}
\arguments{
\item{mean}{the center of the distribution.}

\item{sd}{the dispersion around the center of the distribution.}

\item{nsamples}{(integer) the number of random numbers.}

\item{x}{(array) to evaluate the density function.}

\item{log}{(Boolean) indicates if we are evaluating the \code{log-density}.}
}
\value{
\link{array}

\link{array}
}
\description{
Defines a Normal random variable. As other
objects this is \link[R6:R6Class]{R6} object.

Initializes the random variable

Generates random variables using the \code{rnorm} function.

Evalluates the density function using the \code{pnorm} function.
}
\details{
Should I use this?
}
\section{Fields}{

\describe{
\item{\code{mean}}{(\code{double})\cr
the center of the distribution.}

\item{\code{sd}}{(\code{positive})\cr
the dispersion around the mean.}
}}

\concept{continuous random variables}
