% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{generate.random.samples}
\alias{generate.random.samples}
\title{Generate random samples from a distribution}
\usage{
generate.random.samples(dist, n)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{n}{The number of random samples to be generated}
}
\value{
If n==1, a numeric vector with one value for each variable in the distribution. If there is only one variable in the distribution, a numeric vector with n values. If n>1 and there is more than one variable in the distribution, a matrix with n rows, where each row represents one random sample, and each column represents a variable in the distribution
}
\description{
Generate random samples from a distribution
}
\details{
Note to developers: This function should not be overridden directly. Insteady override \code{\link{do.generate.random.samples}}
}
