% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{get.covariance.matrix}
\alias{get.covariance.matrix}
\title{Get the variance-covariance matrix for variables in a distribution}
\usage{
get.covariance.matrix(dist, n.sim = 1000)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{n.sim}{If the values are to be calculated by random sample generation, how many random samples to generate}
}
\value{
A square matrix of dimension number variables in distribution representing the variance-covariance matrix for the distribution
}
\description{
Get the variance-covariance matrix for variables in a distribution
}
\details{
Note to developers: this function should not be overridden. Override \code{\link{do.get.covariance.matrix}} instead
}
