% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\name{get.defined.transformation}
\alias{get.defined.transformation}
\title{Get Pre-Defined Transformation Objects by Name}
\usage{
get.defined.transformation(transformation.names, throw.error.if.no.match = F)
}
\value{
If transformation.names is a single character value, returns an object of class transformation. Otherwise, returns a list of transformation objects
}
\description{
Gets a pre-defined transformation object. The available transformations are 'log', 'logit', and 'identity'
}
