% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{get.density.function}
\alias{get.density.function}
\title{Get a function that calculates the density for the distribution}
\usage{
get.density.function(dist, default.log = T, n.sim = 1000)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{default.log}{Whether the returned function should calculate the density on the log scale (by default, the returned function can have a log parameter set explicitly)}

\item{n.sim}{If the values are to be calculated by random sample generation, how many random samples to generate}
}
\value{
A function that takes two arguments. The first, mandatory, is 'x' - the values of the parameters at which to calculate the density. May be either a vector, if the density at a single point is desired, or a matrix where each row represents one point and each column represents a variable in the distribution
The second, optional, is 'log' - whether to calculate the density on the log scale. By default, log is set to default.log
}
\description{
A convenience function that packages up a call to \code{\link{calculate.density}} on this distribution into its own stand-alone function (which could be passed along as an argument to other functions)
}
