% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics_definition.R
\name{get.description}
\alias{get.description}
\title{Get the description of a Distribution object}
\usage{
get.description(object)
}
\arguments{
\item{object}{The object to describe}
}
\description{
Get the description of a Distribution object
}
