% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{get.means}
\alias{get.means}
\title{Get means of the variables in a distribution}
\usage{
get.means(dist, n.sim = 1000)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{n.sim}{If the values are to be calculated by random sample generation, how many random samples to generate}
}
\value{
A numeric vector with a mean for each variable in the distribution
}
\description{
Get means of the variables in a distribution
}
\details{
Note to developers: this function should not be overridden. Override \code{\link{do.get.means}} instead
}
