% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{get.medians}
\alias{get.medians}
\title{Get medians of the variables in a distribution}
\usage{
get.medians(dist, n.sim = 1000)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{n.sim}{If the values are to be calculated by random sample generation, how many random samples to generate}
}
\value{
A numeric vector with a median for each variable in the distribution
}
\description{
Get medians of the variables in a distribution
}
\details{
Note to developers: This function should not be overridden directly. Insteady override \code{\link{do.get.medians}}
}
