% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{get.quantiles}
\alias{get.quantiles}
\title{Get quantiles for the variables in a distribution}
\usage{
get.quantiles(dist, p, lower.tail = T, log.p = F, n.sim = 1000)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{p}{A vector of probabilities for which to get the quantiles}

\item{lower.tail}{If true, the returned values are the greatest values x for each variable such that \eqn{P(X \leq x) \geq p}. If fale, the returned values are the smallest values x for each variable such that \eqn{P(X > x) \geq p}. This can be either a scalar (in which case the value is applied to all quantiles), or a vector with one logical value for each value of p}

\item{n.sim}{If the values are to be calculated by random sample generation, how many random samples to generate}
}
\value{
If there is only one variable in the distribution, a vector of quantiles for each value of p. If p is length one, a vector of quantiles for each variable in the distribution. If there is more than one variable in the distribution and p>1, a matrix with one row for each value of p, where each column represents a variable in the distribution
}
\description{
Get quantiles for the variables in a distribution
}
\details{
Note to developers: This function should not be overridden directly. Insteady override \code{\link{do.get.quantiles}}
}
