% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{get.random.generator.function}
\alias{get.random.generator.function}
\title{Get a function that generates random samples form the distribution}
\usage{
get.random.generator.function(dist)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}
}
\value{
A function that takes one argument, 'n' - the number of random samples to generate
}
\description{
A convenience function that packages up a call to \code{\link{generate.random.samples}} on this distribution into its own stand-alone function (which could be passed along as an argument to other functions)
}
