% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{get.sds}
\alias{get.sds}
\title{Get the standard deviations of the variables in a distribution}
\usage{
get.sds(dist, n.sim = 1000)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{n.sim}{If the values are to be calculated by random sample generation, how many random samples to generate}
}
\value{
A numeric vector with a standard deviation for each variable in the distribution
}
\description{
Get the standard deviations of the variables in a distribution
}
\details{
Note to developers: this function should not be overridden. Override \code{\link{do.get.sds}} instead
}
