% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{get.support.bounds}
\alias{get.support.bounds}
\title{Get the maximum and minimm values within a support over real numbers}
\usage{
get.support.bounds(support)
}
\arguments{
\item{support}{An object of class \link{Support}}
}
\value{
If called on an object of class \link{Support}, returns a numeric vector of length 2, with lower and upper bounds. If called on a list containing only \link{Support} objects, returns a matrix with two rows (lower and upper bounds) and a column for each element in the list
}
\description{
Get the maximum and minimm values within a support over real numbers
}
