% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{initialize,Distribution-method}
\alias{initialize,Distribution-method}
\title{A constructor}
\usage{
\S4method{initialize}{Distribution}(
  .Object,
  var.names = NULL,
  n.var = length(var.names),
  support = Multivariate.Support(lapply(1:n.var, function(i) {     Continuous.Support()
    })),
  is.improper = F
)
}
\arguments{
\item{var.names, n.var}{The name and number of variables in the distribution. var.names may be NULL}

\item{subclass}{The name of the subclass to be instantiated with a call to 'new'}

\item{lower.bounds, upper.bounds}{Vectors of lower and upper bounds for each of the variables in the distribution. May be NA}

\item{is.discrete, is.improper}{Logical vectors indicating whether each variable in the distribution is discrete or improper. If a scalar logical is passed, it is assumed to be the value for all variables in the distribution}

\item{...}{Other arguments to be passed to the call to 'new'}
}
\description{
This function is provided for use by developers, and should not be called directly. This function should be called by the subclass constructor. It performs consistency checks on the basic slots for a Distribution object
}
\keyword{internal}
