% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{is.supported}
\alias{is.supported}
\title{Test if numeric values fall within a defined support}
\usage{
is.supported(support, x, by.variable = F)
}
\arguments{
\item{support}{An object of class \link{Support}}

\item{x}{If the support is univariate, a numeric vector. If the support is multivariate, a matrix with one column for each variable and one row for each point to test or a numeric vector with one value for each variable in the support}

\item{by.variable}{If this is a Multivariate_Support, if by.variable is TRUE, calculates support for each variable independently}
}
\value{
If the support is univariate or if by.variable==FALSE or if only a vector is passed to x, returns a logical vector, with one value for each value of x, corresponding to whether that value of x falls within the defined support. If the support is multivariate and by.variable==TRUE and a matrix with more than one row was passed to x, returns a matrix with one column for each variable and one row for each point
}
\description{
Test if numeric values fall within a defined support
}
