% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_independent_distributions.R
\name{join.distributions}
\alias{join.distributions}
\title{Join Multiple Distributions}
\usage{
join.distributions(..., var.names = NULL)
}
\arguments{
\item{...}{One or more Distribution objects or lists of Distribution objects to be joined. If var.names are set on the component distributions, those names will be preserved. If not, and if the elements of ... are named (or elements in lists contained in ... are named), those names will be used as the var.names for the corresponding sub-distributions}
}
\value{
An object of class Joint_Independent_Distributions
}
\description{
Joins multiple Distribution objects into a single Distribution object, on the premise that the variables in each sub-distribution are independent of the variables in any other sub-distribution
}
\seealso{
Other Distribution Constructors: 
\code{\link{Autoregressive.Multivariate.Normal.Distribution}()},
\code{\link{Bernoulli.Distribution}()},
\code{\link{Beta.Distribution}()},
\code{\link{Binomial.Distribution}()},
\code{\link{Canonical.Mixture.Distribution}()},
\code{\link{Compound.Symmetry.Multivariate.Normal.Distribution}()},
\code{\link{Constant.Distribution}()},
\code{\link{Discrete.Set.Distribution}()},
\code{\link{Empiric.Distribution}()},
\code{\link{Logitnormal.Distribution}()},
\code{\link{Logitnormal.Mixture}()},
\code{\link{Logituniform.Distribution}()},
\code{\link{Lognormal.Distribution}()},
\code{\link{Lognormal.Mixture}()},
\code{\link{Loguniform.Distribution}()},
\code{\link{Multivariate.Correlated.Uniform.Distribution}()},
\code{\link{Multivariate.Logitnormal.Distribution}()},
\code{\link{Multivariate.Lognormal.Distribution}()},
\code{\link{Multivariate.Normal.Distribution}()},
\code{\link{Normal.Distribution}()},
\code{\link{Normal.Mixture}()},
\code{\link{Smoothed.Empiric.Distribution}()},
\code{\link{Transformed.Multivariate.Normal.Distribution}()},
\code{\link{Transformed.Normal.Distribution}()},
\code{\link{Transformed.Normal.Mixture}()},
\code{\link{Uniform.Distribution}()},
\code{\link{Univariate.Canonical.Distribution}()}
}
\concept{Distribution Constructors}
