% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joint_independent_distributions.R
\docType{class}
\name{Joint_Independent_Distributions}
\alias{Joint_Independent_Distributions}
\alias{Joint_Independent_Distributions-class}
\title{The Joint_Independent_Distributions class}
\description{
A class that joins multiple independent sub-distributions into a joint distribution
}
\section{Slots}{

\describe{
\item{\code{subdistributions}}{A list of the Distribution objects representing the component distributions of this joint distribution}

\item{\code{indices.for.subdistributions}}{A list whose elements are integer vectors, such that indices.for.subdistributions[[i]] is the indices into the overall distribution of the variables represented by subdistributions[[i]]}

\item{\code{n.subdistributions}}{The number of component distributions}
}}

\seealso{
\link{join.distributions}
}
