% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_canonical_constructors.R
\name{Logituniform.Distribution}
\alias{Logituniform.Distribution}
\title{Create a Logit-Uniform Distribution Object}
\usage{
Logituniform.Distribution(min = 0, max = 1, var.name = NULL)
}
\arguments{
\item{min, max}{The bounds of the distribution (on the original scale, NOT the logit scale). If min is 0 or max is 1, returns an improper logit-uniform distribution (which returns 1/x/(1-x) for all densities, but cannot evaluate cdfs, quantiles, or generate random samples)}

\item{var.name}{The name of the single variable in the distribution, or NULL if no name is specified}
}
\description{
Create a Logit-Uniform Distribution Object
}
\seealso{
Other Univariate Canonical Distribution Constructors: 
\code{\link{Bernoulli.Distribution}()},
\code{\link{Beta.Distribution}()},
\code{\link{Binomial.Distribution}()},
\code{\link{Canonical.Mixture.Distribution}()},
\code{\link{Logitnormal.Distribution}()},
\code{\link{Lognormal.Distribution}()},
\code{\link{Loguniform.Distribution}()},
\code{\link{Normal.Distribution}()},
\code{\link{Transformed.Normal.Distribution}()},
\code{\link{Uniform.Distribution}()},
\code{\link{Univariate.Canonical.Distribution}()}

Other Distribution Constructors: 
\code{\link{Autoregressive.Multivariate.Normal.Distribution}()},
\code{\link{Bernoulli.Distribution}()},
\code{\link{Beta.Distribution}()},
\code{\link{Binomial.Distribution}()},
\code{\link{Canonical.Mixture.Distribution}()},
\code{\link{Compound.Symmetry.Multivariate.Normal.Distribution}()},
\code{\link{Empiric.Distribution}()},
\code{\link{Logitnormal.Distribution}()},
\code{\link{Logitnormal.Mixture}()},
\code{\link{Lognormal.Distribution}()},
\code{\link{Lognormal.Mixture}()},
\code{\link{Loguniform.Distribution}()},
\code{\link{Multivariate.Logitnormal.Distribution}()},
\code{\link{Multivariate.Lognormal.Distribution}()},
\code{\link{Multivariate.Normal.Distribution}()},
\code{\link{Normal.Distribution}()},
\code{\link{Normal.Mixture}()},
\code{\link{Smoothed.Empiric.Distribution}()},
\code{\link{Transformed.Multivariate.Normal.Distribution}()},
\code{\link{Transformed.Normal.Distribution}()},
\code{\link{Transformed.Normal.Mixture}()},
\code{\link{Uniform.Distribution}()},
\code{\link{Univariate.Canonical.Distribution}()},
\code{\link{join.distributions}()}
}
\concept{Distribution Constructors}
\concept{Univariate Canonical Distribution Constructors}
