% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_canonical_constructors.R
\name{Lognormal.Distribution}
\alias{Lognormal.Distribution}
\title{Create a Univariate Log-Normal Distribution Object}
\usage{
Lognormal.Distribution(
  meanlog = 0,
  sdlog = 1,
  lower = 0,
  upper = Inf,
  var.name = NULL
)
}
\arguments{
\item{meanlog, sdlog}{The mean and standard deviation (on the log scale) of the distribution}

\item{lower, upper}{The lower and the upper bounds (on the original scale, NOT the log scale), if this is a bounded distribution}

\item{var.name}{The name of the single variable in the distribution, or NULL if no name is specified}
}
\description{
Create a Univariate Log-Normal Distribution Object
}
\seealso{
Other Univariate Canonical Distribution Constructors: 
\code{\link{Bernoulli.Distribution}()},
\code{\link{Beta.Distribution}()},
\code{\link{Binomial.Distribution}()},
\code{\link{Canonical.Mixture.Distribution}()},
\code{\link{Logitnormal.Distribution}()},
\code{\link{Logituniform.Distribution}()},
\code{\link{Loguniform.Distribution}()},
\code{\link{Normal.Distribution}()},
\code{\link{Transformed.Normal.Distribution}()},
\code{\link{Uniform.Distribution}()},
\code{\link{Univariate.Canonical.Distribution}()}

Other Distribution Constructors: 
\code{\link{Autoregressive.Multivariate.Normal.Distribution}()},
\code{\link{Bernoulli.Distribution}()},
\code{\link{Beta.Distribution}()},
\code{\link{Binomial.Distribution}()},
\code{\link{Canonical.Mixture.Distribution}()},
\code{\link{Compound.Symmetry.Multivariate.Normal.Distribution}()},
\code{\link{Empiric.Distribution}()},
\code{\link{Logitnormal.Distribution}()},
\code{\link{Logitnormal.Mixture}()},
\code{\link{Logituniform.Distribution}()},
\code{\link{Lognormal.Mixture}()},
\code{\link{Loguniform.Distribution}()},
\code{\link{Multivariate.Logitnormal.Distribution}()},
\code{\link{Multivariate.Lognormal.Distribution}()},
\code{\link{Multivariate.Normal.Distribution}()},
\code{\link{Normal.Distribution}()},
\code{\link{Normal.Mixture}()},
\code{\link{Smoothed.Empiric.Distribution}()},
\code{\link{Transformed.Multivariate.Normal.Distribution}()},
\code{\link{Transformed.Normal.Distribution}()},
\code{\link{Transformed.Normal.Mixture}()},
\code{\link{Uniform.Distribution}()},
\code{\link{Univariate.Canonical.Distribution}()},
\code{\link{join.distributions}()}
}
\concept{Distribution Constructors}
\concept{Univariate Canonical Distribution Constructors}
