% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_constructors.R
\name{Multivariate.Correlated.Uniform.Distribution}
\alias{Multivariate.Correlated.Uniform.Distribution}
\title{Create a Multivariate Uniform Distribution Object Where the Variables are Correlated}
\usage{
Multivariate.Correlated.Uniform.Distribution(
  correlation.matrix,
  min = rep(0, dim(correlation.matrix)[1]),
  max = rep(1, dim(correlation.matrix)[1]),
  var.names = NULL
)
}
\arguments{
\item{correlation.matrix}{The matrix specifying how variables should be correlated}

\item{min, max}{The vectors of minima and maxima for each variable in the distribution. If given a scalar value, assumes that value applies for all variables}

\item{var.names}{The names of the variables in the distribution. If NULL, uses the names of mu, the row names of sigma, or the column names of sigma, in that order}
}
\description{
Uses a rough approximation to generate a multivariate normal distribution such that the correlations between variables are approximately equal to the given correlation.matrix. Specifically, generates random samples from a multivariate normal distribution with mean=0 and covariance matrix = cov2cor(correlation.matrix + 0.04*I)
}
\seealso{
Other Distribution Constructors: 
\code{\link{Autoregressive.Multivariate.Normal.Distribution}()},
\code{\link{Bernoulli.Distribution}()},
\code{\link{Beta.Distribution}()},
\code{\link{Binomial.Distribution}()},
\code{\link{Canonical.Mixture.Distribution}()},
\code{\link{Compound.Symmetry.Multivariate.Normal.Distribution}()},
\code{\link{Constant.Distribution}()},
\code{\link{Discrete.Set.Distribution}()},
\code{\link{Empiric.Distribution}()},
\code{\link{Logitnormal.Distribution}()},
\code{\link{Logitnormal.Mixture}()},
\code{\link{Logituniform.Distribution}()},
\code{\link{Lognormal.Distribution}()},
\code{\link{Lognormal.Mixture}()},
\code{\link{Loguniform.Distribution}()},
\code{\link{Multivariate.Logitnormal.Distribution}()},
\code{\link{Multivariate.Lognormal.Distribution}()},
\code{\link{Multivariate.Normal.Distribution}()},
\code{\link{Normal.Distribution}()},
\code{\link{Normal.Mixture}()},
\code{\link{Smoothed.Empiric.Distribution}()},
\code{\link{Transformed.Multivariate.Normal.Distribution}()},
\code{\link{Transformed.Normal.Distribution}()},
\code{\link{Transformed.Normal.Mixture}()},
\code{\link{Uniform.Distribution}()},
\code{\link{Univariate.Canonical.Distribution}()},
\code{\link{join.distributions}()}
}
\concept{Distribution Constructors}
