% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_constructors.R
\name{Multivariate.Logitnormal.Distribution}
\alias{Multivariate.Logitnormal.Distribution}
\title{Create a Multivariate Logit-Normal Distribution Object}
\usage{
Multivariate.Logitnormal.Distribution(
  mu = rep(0, nrow(sigma)),
  sigma = diag(rep(1, length(mu))),
  lower = 0,
  upper = Inf,
  var.names = NULL
)
}
\arguments{
\item{mu, sigma}{The mean vector and covariance matrix of the distribution (after logit transformation)}

\item{lower, upper}{The upper and lower bounds for each variable (on the original scale, NOT the logit scale). If only one value is given, assumed to apply to all variables in the distribution}

\item{var.names}{The names of the variables in the distribution. If NULL, uses the names of mu, the row names of sigma, or the column names of sigma, in that order}
}
\description{
Create a Multivariate Logit-Normal Distribution Object
}
\seealso{
Other Distribution Constructors: 
\code{\link{Autoregressive.Multivariate.Normal.Distribution}()},
\code{\link{Bernoulli.Distribution}()},
\code{\link{Beta.Distribution}()},
\code{\link{Binomial.Distribution}()},
\code{\link{Canonical.Mixture.Distribution}()},
\code{\link{Compound.Symmetry.Multivariate.Normal.Distribution}()},
\code{\link{Empiric.Distribution}()},
\code{\link{Logitnormal.Distribution}()},
\code{\link{Logitnormal.Mixture}()},
\code{\link{Logituniform.Distribution}()},
\code{\link{Lognormal.Distribution}()},
\code{\link{Lognormal.Mixture}()},
\code{\link{Loguniform.Distribution}()},
\code{\link{Multivariate.Correlated.Uniform.Distribution}()},
\code{\link{Multivariate.Lognormal.Distribution}()},
\code{\link{Multivariate.Normal.Distribution}()},
\code{\link{Normal.Distribution}()},
\code{\link{Normal.Mixture}()},
\code{\link{Smoothed.Empiric.Distribution}()},
\code{\link{Transformed.Multivariate.Normal.Distribution}()},
\code{\link{Transformed.Normal.Distribution}()},
\code{\link{Transformed.Normal.Mixture}()},
\code{\link{Uniform.Distribution}()},
\code{\link{Univariate.Canonical.Distribution}()},
\code{\link{join.distributions}()}
}
\concept{Distribution Constructors}
