% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{Multivariate.Support}
\alias{Multivariate.Support}
\title{Join multiple Support objects into a multivariate Support}
\usage{
Multivariate.Support(...)
}
\arguments{
\item{...}{Objects of class \link{Support}, or lists that contain only objects of class \link{Support}}
}
\value{
A \link{Multivariate_Support} object
}
\description{
Join multiple Support objects into a multivariate Support
}
