% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariate_normal_distribution.R
\docType{class}
\name{Multivariate_Normal_Distribution}
\alias{Multivariate_Normal_Distribution}
\alias{Multivariate_Normal_Distribution-class}
\title{The Multivariate_Normal_Distribution Class}
\description{
Defines a multivariate normal distribution
}
\section{Slots}{

\describe{
\item{\code{mu,sigma}}{The mean vector and covariance matrix of the distribution}
}}

