% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\docType{class}
\name{Multivariate_Support}
\alias{Multivariate_Support}
\alias{Multivariate_Support-class}
\title{An object representing support (on the real numbers) over across multiple variables}
\description{
An object representing support (on the real numbers) over across multiple variables
}
