% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete.R
\name{PoissonRandomVariable}
\alias{PoissonRandomVariable}
\title{PoissonRandomVariable object}
\description{
Defines a Poisson random variable. It uses in the backend
all functions that defines a \code{\link[=rpois]{Poisson}} random variable. As
the other definitions in \code{distributions} this is an
\code{\link[R6:R6Class]{R6}} object. Tha parametrization used
is from a \deqn{X \sim \mathsf{Poisson}(\lambda)\,,}
where the parameter \eqn{\lambda >0} denotes the rate of ocurrence
of events.
\subsection{Note}{

All random variables in \code{distributions} are defined as \code{\link{R6}}
objects. The \code{fields} (referenced below) are needed for objects of
\code{\link[R6:R6Class]{R6::R6Class}}. In our context these are what defines
which specific instance of a
random variable is used. That is, the public \code{fields} are the
parameters of the random variable \eqn{\theta \in \mathbb{R}^p}, for some
\eqn{p}.
}
}
\seealso{
Other discrete: 
\code{\link{BionomialRandomVariable}}
}
\concept{discrete}
\section{Super class}{
\code{distributions::DiscreteRandomVariable} -> \code{Poisson}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{lambda}}{(\code{double}, positive) \cr
the number of independent trials, \eqn{\lambda > 0}.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Poisson-new}{\code{PoissonRandomVariable$new()}}
\item \href{#method-Poisson-sample}{\code{PoissonRandomVariable$sample()}}
\item \href{#method-Poisson-density}{\code{PoissonRandomVariable$density()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poisson-new"></a>}}
\if{latex}{\out{\hypertarget{method-Poisson-new}{}}}
\subsection{Method \code{new()}}{
Generates a Binomial \code{\link{DiscreteRandomVariable}} object with specified
number of trials (\code{n}) and probability of success (\code{theta}).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PoissonRandomVariable$new(lambda = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{lambda}}{(\code{double}, positive) the rate of events.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poisson-sample"></a>}}
\if{latex}{\out{\hypertarget{method-Poisson-sample}{}}}
\subsection{Method \code{sample()}}{
Generates random variables using the \code{\link[stats:Poisson]{stats::rpois()}} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PoissonRandomVariable$sample(nsamples = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{nsamples}}{(integer) the number of random numbers
being generated that behaves like the random variables.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{array}) of random numbers from the distribution.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Poisson-density"></a>}}
\if{latex}{\out{\hypertarget{method-Poisson-density}{}}}
\subsection{Method \code{density()}}{
Evaluates the density function using the \code{\link[stats:Poisson]{stats::dpois()}} function.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{PoissonRandomVariable$density(x, log = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(array) to evaluate the density function.}

\item{\code{log}}{(Boolean) indicates if we are evaluating the \code{log-density}.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
(\code{array}) of evaluations at \code{x}.
}
}
}
