% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{random}
\alias{random}
\title{Create random samples from a normal distribution}
\usage{
random(nsamples)
}
\arguments{
\item{nsamples}{an integer}
}
\value{
A vector with random numbers
}
\description{
Create random samples from a normal distribution
}
\examples{
x <- random(10)
}
