% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\docType{class}
\name{Range_Support}
\alias{Range_Support}
\alias{Range_Support-class}
\title{An object representing support over a range of real numbers}
\description{
An object representing support over a range of real numbers
}
