% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{separate.independent.distributions}
\alias{separate.independent.distributions}
\title{Separate a distribution object into a list of distributions that are independent from each other}
\usage{
separate.independent.distributions(object, var.names)
}
\arguments{
\item{object}{A distribution object}
}
\value{
A list of Distribution objects, where each distribution either (a) is univariate or (b) contains only correlated variables
}
\description{
Separate a distribution object into a list of distributions that are independent from each other
}
