% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{set.var.names}
\alias{set.var.names}
\title{Sets the variable names for a Distribution Object}
\usage{
set.var.names(object, var.names)
}
\arguments{
\item{var.names}{The character vector of variable names}
}
\value{
A Distribution object with the new var.names set
}
\description{
Sets the variable names for a Distribution Object
}
