% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoothed_empiric_distribution.R
\docType{class}
\name{Smoothed_Empiric_Distribution}
\alias{Smoothed_Empiric_Distribution}
\alias{Smoothed_Empiric_Distribution-class}
\title{The Smoothed_Empiric_Distribution Class}
\description{
A class that defines a distribution based on a set of samples from that distribution, but applies Gaussian kernel smoothers. For an unsmoothed representation, see \link{Empiric_Distribution}
}
\section{Slots}{

\describe{
\item{\code{points}}{A matrix containing the samples}

\item{\code{weights}}{A numeric vector with the weight given to each row in points. Sum to 1}

\item{\code{n.points}}{The number of rows in points}
}}

\seealso{
\link{Empiric_Distribution}, \link{Smoothed.Empiric.Distribution}
}
