% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master_class.R
\name{subset.distribution}
\alias{subset.distribution}
\title{Get a distribution that describes a subset of the variables in a given distribution}
\usage{
subset.distribution(dist, vars.to.keep)
}
\arguments{
\item{dist}{An object of class Distribution or a subclass}

\item{vars.to.keep}{Either a character vector with the names of the variables to keep, an integer/numeric vector with the indices of the variables to keep, or a logical vector of length n.var indicating which variables to keep}
}
\value{
An object of class \code{\link{Distribution-class}}
}
\description{
Get a distribution that describes a subset of the variables in a given distribution
}
\details{
Note to developers: this function should not be overridden. Override \code{\link{do.subset.distribution}} instead
}
