% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{subset.support}
\alias{subset.support}
\title{Subset a Support object}
\usage{
subset.support(support, keep)
}
\arguments{
\item{support}{An object of class \link{Support}}

\item{keep}{Which dimensions of the support to keep. Either a set of integer indices or a logical vector}
}
\value{
A \link{Support} object
}
\description{
Subset a Support object
}
