% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\name{Support.Union}
\alias{Support.Union}
\title{Join multiple univariate Support objects into one univariate Support}
\usage{
Support.Union(...)
}
\arguments{
\item{...}{Objects of class \link{Support}, or lists that contain only objects of class \link{Support}}
}
\value{
A \link{Union_Support} object, such that is.supported is true if is.supported is true for any of its component supports
}
\description{
Join multiple univariate Support objects into one univariate Support
}
