% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformations.R
\docType{class}
\name{transformation}
\alias{transformation}
\alias{transformation-class}
\title{The transformation class}
\description{
A class that represents a functional transformation, along with how to reverse the transformation and take its derivative
}
\section{Slots}{

\describe{
\item{\code{transform.function,reverse.transform.function,transformation.derivative,log.abs.transformation.derivative}}{Functions that compute the transformation, inverse transformation, derivative of the transformation function, and log of the absolute value of the derivative of the transformation function, directly. All should take one argument, 'x', which can be either a numeric vector or numeric scalar. log.abs.transformation.derivative may be supplied as NULL; the other three functions must be defined}

\item{\code{name}}{A character value with a descriptive name of the transformation}
}}

