% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_mixture_constructors.R
\name{Transformed.Normal.Mixture}
\alias{Transformed.Normal.Mixture}
\title{Create a Mixture of Transformed Normal Distributions}
\usage{
Transformed.Normal.Mixture(
  means = 0,
  sds = 1,
  weights = 1,
  transformation = NULL,
  lower = -Inf,
  upper = Inf,
  var.name = NULL
)
}
\arguments{
\item{means, sds}{The means and standard deviations of each component of the mixture (after transformation has been applied)}

\item{weights}{The weights applied to each component of the mixture. Weights need not be normalized to sum to 1. If a scalar value is given, all components of the mixture will have the same weight}

\item{lower, upper}{The lower and the upper bounds, if this is a bounded distribution. The bounds are on the ORIGINAL scale (before transformation). The same bounds apply for all components of the mixture}
}
\description{
Creates a mixture of distributions representing random variables that are normal after transformation (eg, mixture of normals, log-normals, logit-normals, etc.)
}
\seealso{
Other Canonical Mixture Distribution Constructors: 
\code{\link{Logitnormal.Mixture}()},
\code{\link{Lognormal.Mixture}()},
\code{\link{Normal.Mixture}()}

Other Distribution Constructors: 
\code{\link{Autoregressive.Multivariate.Normal.Distribution}()},
\code{\link{Bernoulli.Distribution}()},
\code{\link{Beta.Distribution}()},
\code{\link{Binomial.Distribution}()},
\code{\link{Canonical.Mixture.Distribution}()},
\code{\link{Compound.Symmetry.Multivariate.Normal.Distribution}()},
\code{\link{Empiric.Distribution}()},
\code{\link{Logitnormal.Distribution}()},
\code{\link{Logitnormal.Mixture}()},
\code{\link{Logituniform.Distribution}()},
\code{\link{Lognormal.Distribution}()},
\code{\link{Lognormal.Mixture}()},
\code{\link{Loguniform.Distribution}()},
\code{\link{Multivariate.Logitnormal.Distribution}()},
\code{\link{Multivariate.Lognormal.Distribution}()},
\code{\link{Multivariate.Normal.Distribution}()},
\code{\link{Normal.Distribution}()},
\code{\link{Normal.Mixture}()},
\code{\link{Smoothed.Empiric.Distribution}()},
\code{\link{Transformed.Multivariate.Normal.Distribution}()},
\code{\link{Transformed.Normal.Distribution}()},
\code{\link{Uniform.Distribution}()},
\code{\link{Univariate.Canonical.Distribution}()},
\code{\link{join.distributions}()}
}
\concept{Canonical Mixture Distribution Constructors}
\concept{Distribution Constructors}
