% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/support.R
\docType{class}
\name{Union_Support}
\alias{Union_Support}
\alias{Union_Support-class}
\title{An object representing the union of two or more supports over a range of real numbers}
\description{
An object representing the union of two or more supports over a range of real numbers
}
