% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_canonical_distribution.R
\name{Univariate.Canonical.Distribution}
\alias{Univariate.Canonical.Distribution}
\title{Create a distribution object based off of a canonical, univariate distribution}
\usage{
Univariate.Canonical.Distribution(
  name,
  dist.name,
  parameters,
  var.name = NULL,
  support = Continuous.Support(),
  density.function.name = paste0("d", dist.name),
  cdf.function.name = paste0("p", dist.name),
  quantile.function.name = paste0("q", dist.name),
  transformation = NULL,
  is.improper = F,
  is.discrete = F,
  mean.value = as.numeric(NA),
  variance.value = as.numeric(NA)
)
}
\arguments{
\item{name}{A short, descriptive name for the distribution}

\item{dist.name}{The root name of the distribution as used by d/r/p functions in R. For example, to create a normal distribution, dist.name='norm' (like dnorm, rnorm, pnorm). To create a beta distribution, dist.name='beta' (dbeta, rbeta, qbeta). This argument is only necessary if density.function, cdf.function, and quantile.function are not passed explicitly}

\item{parameters}{A list of parameters that would be passed to d/r/p functions in R. For example, for a normal(0,1), parameters=list(mean=0, sd=1)}

\item{var.name}{The name of the variable in this distribution}

\item{support}{A \link{support} object that}

\item{transformation}{A transformation object, if the named distribution operates on a transformation of the random variable (can also pass NULL for no transformation or the name of a predefined transformation - see \code{\link{get.defined.transformation}}). For example, to create a log-normal distribution, use dist.name='norm' with transformation='log}

\item{is.improper, is.discrete}{Logicals indicating whether the distribution is improper/discrete}

\item{mean.value, variance}{If known, these values will be returned for calls to get.means, get.sds, get.covariance.matrix. If passed NA, random sampling will be used to estimate distribution parameters if requested}

\item{density.function, cdf.function, quantile.function}{The functions that calculate the density, cdf, and quantile respectively. These should take the standard parameters of such functions in R. density.function should take 'x', distribution-specific parameters, and 'log'. cdf.function should take 'q', distribution-specific parameters, 'lower.tail' and 'log.p'. quantile.function should take 'p', distribution-specific parameters, and 'lower.tail' and 'log.p'}
}
\description{

}
\seealso{
\code{\link{get.defined.transformation}}, \code{\link{create.transformation}}

Other Univariate Canonical Distribution Constructors: 
\code{\link{Bernoulli.Distribution}()},
\code{\link{Beta.Distribution}()},
\code{\link{Binomial.Distribution}()},
\code{\link{Canonical.Mixture.Distribution}()},
\code{\link{Logitnormal.Distribution}()},
\code{\link{Logituniform.Distribution}()},
\code{\link{Lognormal.Distribution}()},
\code{\link{Loguniform.Distribution}()},
\code{\link{Normal.Distribution}()},
\code{\link{Transformed.Normal.Distribution}()},
\code{\link{Uniform.Distribution}()}

Other Distribution Constructors: 
\code{\link{Autoregressive.Multivariate.Normal.Distribution}()},
\code{\link{Bernoulli.Distribution}()},
\code{\link{Beta.Distribution}()},
\code{\link{Binomial.Distribution}()},
\code{\link{Canonical.Mixture.Distribution}()},
\code{\link{Compound.Symmetry.Multivariate.Normal.Distribution}()},
\code{\link{Empiric.Distribution}()},
\code{\link{Logitnormal.Distribution}()},
\code{\link{Logitnormal.Mixture}()},
\code{\link{Logituniform.Distribution}()},
\code{\link{Lognormal.Distribution}()},
\code{\link{Lognormal.Mixture}()},
\code{\link{Loguniform.Distribution}()},
\code{\link{Multivariate.Correlated.Uniform.Distribution}()},
\code{\link{Multivariate.Logitnormal.Distribution}()},
\code{\link{Multivariate.Lognormal.Distribution}()},
\code{\link{Multivariate.Normal.Distribution}()},
\code{\link{Normal.Distribution}()},
\code{\link{Normal.Mixture}()},
\code{\link{Smoothed.Empiric.Distribution}()},
\code{\link{Transformed.Multivariate.Normal.Distribution}()},
\code{\link{Transformed.Normal.Distribution}()},
\code{\link{Transformed.Normal.Mixture}()},
\code{\link{Uniform.Distribution}()},
\code{\link{join.distributions}()}
}
\concept{Distribution Constructors}
\concept{Univariate Canonical Distribution Constructors}
