% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate_canonical_distribution.R
\docType{class}
\name{Univariate_Canonical_Distribution}
\alias{Univariate_Canonical_Distribution}
\alias{Univariate_Canonical_Distribution-class}
\title{The Univariate_Canonical_Distribution class}
\description{
A class that represents a univariate canonical distribution, such as a normal or beta distribution
}
\section{Slots}{

\describe{
\item{\code{name}}{A short, descriptive name for the distribution}

\item{\code{density.function.name,cdf.function.name,quantile.function.name}}{Names of the (elsewhere defined) functions that calculate density, cdf, and quantiles (eg 'dnorm', 'pnorm', and 'qnorm)}

\item{\code{parameters}}{A list of parameters that would be passed to density.function, cdf.function, and quantile.function. For example, for a normal(0,1), parameters=list(mean=0, sd=1)}

\item{\code{p.in.bounds,log.p.in.bounds}}{The cumulative probability and of the log cumulative probability between the lower and upper bounds}

\item{\code{p.lower.bound}}{The value of the cdf evaluated at the lower bound}

\item{\code{transformation}}{An object of class \link{transformation}, defining the transformation applied to the variable before applying the distribution (eg, )}

\item{\code{mean.value,variance}}{If known, the mean and variance of the distribution. May be NA}
}}

\seealso{
The family of functions that create specific distributions, such as \link{Normal.Distribution}
}
