\name{brgen}
\alias{brgen}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Bernoulli Numbers
}
\description{
Brgen function using cugen function to create bernoulli number with parameter p
}
\usage{
brgen(p)
}
\arguments{
  \item{p}{
    \code{
    Probability of success
    }
}
}
\examples{
brgen(0.5)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
