\name{cugen}
\alias{cugen}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Uniform Distribution
}
\description{
This function uses dugen function with parameters 0 and 1 as input and makes uniform random number between  0 and 1.
}
\usage{
cugen()
}
cugen()
\value{
cugen returns a uniform random number between 0 and 1.
}
\examples{
cugen()
## returns random number
}
