% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/continuous.R
\name{GaussianRandomVariable}
\alias{GaussianRandomVariable}
\title{Gaussian random variable object}
\value{
A \code{\link{RandomVariable}} object.
}
\description{
Normal random variable
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mean}}{(double)\cr
the center of the distribution.}

\item{\code{sd}}{(positive)\cr
the dispersion around the mean.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RandomVariable-new}{\code{GaussianRandomVariable$new()}}
\item \href{#method-RandomVariable-sample}{\code{GaussianRandomVariable$sample()}}
\item \href{#method-RandomVariable-density}{\code{GaussianRandomVariable$density()}}
\item \href{#method-RandomVariable-clone}{\code{GaussianRandomVariable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RandomVariable-new"></a>}}
\if{latex}{\out{\hypertarget{method-RandomVariable-new}{}}}
\subsection{Method \code{new()}}{
Initializes the random variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianRandomVariable$new(mean = 0, sd = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mean}}{the center of the distribution.}

\item{\code{sd}}{the dispersion around the center of the distribution.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RandomVariable-sample"></a>}}
\if{latex}{\out{\hypertarget{method-RandomVariable-sample}{}}}
\subsection{Method \code{sample()}}{
Sampler
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianRandomVariable$sample(n = 1)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{n}}{(integer) the number of random numbers.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RandomVariable-density"></a>}}
\if{latex}{\out{\hypertarget{method-RandomVariable-density}{}}}
\subsection{Method \code{density()}}{
Density function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianRandomVariable$density(x, log = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{(array) to evaluate the density function.}

\item{\code{log}}{(Boolean) indicates if we are evaluating the \code{log-density}.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RandomVariable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RandomVariable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{GaussianRandomVariable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
