\name{geometcdf}

\alias{geometcdf}

\title{ Densità cumulata geometrica }

\description{
 
}
\usage{
geometcdf(p, x)
}

\arguments{
  \item{p}{ probabilità }
  \item{x}{ numero della prova in cui si ottiene il primo esito positivo }
}

\references{ \url{http://www.education.ti.com} }

\author{ Fabio Frascati }

\seealso{ \code{\link{geometpdf}} }

\examples{
geometcdf(p=0.5,x=2)}
}

\keyword{ distribution }
